+++ packages/web-speech-cognitive-services/src/SpeechServices/SpeechToText.ts
@@ -1,5 +1,1 @@
-import { createSpeechRecognitionPonyfill, createSpeechRecognitionPonyfillFromRecognizer } from './SpeechToText/index';
-
-export default createSpeechRecognitionPonyfill;
-
-export { createSpeechRecognitionPonyfillFromRecognizer };
+export * from './SpeechToText/index';