+++ packages/web-speech-cognitive-services/src/SpeechServices/SpeechToText/createSpeechRecognitionPonyfill.ts
@@ -26,11 +26,9 @@ export default function createSpeechRecognitionPonyfill(options: PatchOptionsIni
   } = patchOptions(options);
 
   if (!audioConfig && (!window.navigator.mediaDevices || !window.navigator.mediaDevices.getUserMedia)) {
-    console.warn(
-      'web-speech-cognitive-services: This browser does not support WebRTC and it will not work with Cognitive Services Speech Services.'
+    throw new Error(
+      'web-speech-cognitive-services: This browser does not support Media Capture and Streams API and it will not work with Cognitive Services Speech Services.'
     );
-
-    return {};
   }
 
   const createRecognizer = async (lang: string) => {